pyproject.toml
[tool.poetry]
name = "filesff"
version = "0.0.7"
description = "Files for Fun; Python Utilities"
authors = ["Netanel Revah <netanelrevah@outlook.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ujson = { version = "^5.6.0", optional = true }
protobuf = {version = "^4.22.1", optional = true}
cap = {version = "^0.1.4", optional = true}
msgpack = {version = "^1.0.5", optional = true}
boto3 = {version = "^1.26.91", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-tstcls = { git = "https://github.com/netanelrevah/tstcls" }
pytest-parametrization = "^2022.2.1"
mypy = ">=0.991,<1.1"
black = "*"
isort = "*"
ruff = ">=0.0.247,<0.0.255"
msgpack-types = "^0.2.0"
types-protobuf = "^4.21.0.3"
types-ujson = "^5.7.0.0"
boto3-stubs = { version="*", extras=["s3"]}

[tool.poetry.extras]
ujson = ["ujson"]
protobuf = ["protobuf"]
cap = ["cap"]
msgpack = ["msgpack"]
s3 = ["boto3"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 120

[tool.mypy]
[[tool.mypy.overrides]]
module = "cap"
ignore_missing_imports = true